Split JobPool declarations and definitions
