+++ src/openrct2/core/FileIndex.hpp
@@ -14,7 +14,7 @@
 #include "File.h"
 #include "FileScanner.h"
 #include "FileStream.h"
-#include "JobPool.hpp"
+#include "JobPool.h"
 #include "Path.hpp"
 
 #include <chrono>
