+++ src/openrct2/interface/Viewport.cpp
@@ -15,7 +15,7 @@
 #include "../OpenRCT2.h"
 #include "../config/Config.h"
 #include "../core/Guard.hpp"
-#include "../core/JobPool.hpp"
+#include "../core/JobPool.h"
 #include "../drawing/Drawing.h"
 #include "../drawing/IDrawingEngine.h"
 #include "../paint/Paint.h"
